huge update

+++ Informatyka/Klasa 2/najbardziejskomplikowanyprojektwdziejachludzkosci.py
@@ -5,8 +5,8 @@ srednia = 0
 sredniaout = 0
 dzielnik = 0
 for _ in range(1,a+1):
-    b = int(input("jaka ocena?"))
-    c = int(input("jaka waga?"))
+    b = float(input("jaka ocena?"))
+    c = float(input("jaka waga?"))
     srednia += b*c
     dzielnik += c
     sredniaout = srednia / dzielnik
